Move new XboxController class to *.util package
