+++ src/org/usfirst/frc/team88/robot/util/XboxController.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team88.robot;
+package org.usfirst.frc.team88.robot.util;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.Button;
